artifacts/sdk_fuzzer/requirements.txt
fuzzingbook==0.9.5
azure.storage.blob==12.15.0
azure.storage.queue==12.5.0
azure.data.tables==12.4.2
localstack==2.0.1
boto3==1.25.5